02 rtsp_cam_basic/requirements.txt
opencv-python-headless
python-dotenv
onvif-zeep
